Ref: ngx-imgage-gallery move to src
