+++ src/components/ngx-image-gallery/ngx-image-gallery.component.scss
