+++ src/ngx-image-gallery.conf.ts
